Factor out inline script
